Fix compiler warnings on Windows
Warnings were:

  GameInfoLoader.cpp(56): warning C4244: 'initializing': conversion from 'uint64_t' to 'size_t', possible loss of data

  GameInfoLoader.cpp(63): warning C4244: 'argument': conversion from 'int64_t' to 'size_t', possible loss of data

  GameInfoLoader.cpp(77): warning C4244: '=': conversion from 'ssize_t' to 'unsigned int', possible loss of data

  Cheevos.cpp(236): warning C4018: '<': signed/unsigned mismatch

  CheevosFrontendBridge.cpp(57): warning C4244: 'return': conversion from 'const int64_t' to 'size_t', possible loss of data

  InputManager.cpp(132): warning C4018: '>=': signed/unsigned mismatch

  InputManager.cpp(156): warning C4018: '<': signed/unsigned mismatch

  InputManager.cpp(185): warning C4018: '<': signed/unsigned mismatch

  InputManager.cpp(247): warning C4018: '>=': signed/unsigned mismatch

+++ src/GameInfoLoader.cpp
@@ -53,14 +53,14 @@ bool CGameInfoLoader::Load(void)
     return false;
   }
 
-  int64_t size = statStruct.GetSize();
+  uint64_t size = statStruct.GetSize();
   if (size > 0)
   {
     // Size is known, read entire file at once (unless it is too big)
     if (size <= MAX_READ_SIZE)
     {
-      m_dataBuffer.resize((size_t)size);
-      file.Read(m_dataBuffer.data(), size);
+      m_dataBuffer.resize(static_cast<size_t>(size));
+      file.Read(m_dataBuffer.data(), static_cast<size_t>(size));
     }
     else
     {
@@ -72,7 +72,7 @@ bool CGameInfoLoader::Load(void)
   else
   {
     // Read file in chunks
-    unsigned int bytesRead;
+    ssize_t bytesRead;
     uint8_t buffer[READ_SIZE];
     while ((bytesRead = file.Read(buffer, sizeof(buffer))) > 0)
     {
